Robinhood_Bot_DeepSeek/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
robin-stocks>=2.0.3
pyyaml>=5.4.1
pytest>=6.2.5
qmcpy>=1.3.0
scipy>=1.7.0